The Biggest Bug fixed.
The line 212 to 216 caused the whole prediction fail.
I supposed to cut the positives and negatives, and made them balance. but my the line 214 & 215 in the code was wrongly wrote.
As the type of label is FLOAT, I compared them with STRING ('1.0' and '0.0'), so the training data was built wrongly.
There are no need to change any others code but line 212 to 216, the AUCs can raise from 0.545 to 0.660(Decision Tree), 0.742(Random Forest) and 0.681(gradient boost decision tree).
The code is executable, all need to do is modified the input data(@ line 33 to 37) and output path(@ line 39).